+++ lib/src/move/move_list_extension.dart
@@ -16,6 +16,9 @@ extension MoveListExtension<T extends Move> on Iterable<T> {
   /// All moves that involve a capture.
   List<Move> get captures => where((e) => e.capture).toList();
 
+  /// All moves that don't involve a capture.
+  List<Move> get quiet => where((e) => !e.capture).toList();
+
   /// Moves that are a pass. Usually should only be 0 or 1 of these.
   List<PassMove> get passMoves =>
       where((e) => e is PassMove).map((e) => e as PassMove).toList();