+++ CHANGELOG.md
@@ -9,6 +9,9 @@
   - Legan Chess (`MiscVariants.legan()`): a diagonally arranged chess variant.
   - Grasshopper Chess (`FairyVariants.grasshopper()`): chess with some extra grasshopper pieces.
   - Berolina Chess (`FairyVariants.berolina()`): chess with the Berolina pawn (`PieceType.berolinaPawn()`).
+- A lot of old deprecated constants and functions have been removed.
+- `play.dart` example: move filtering commands - `moves from x`, `moves to x`, `moves captures`, `moves quiet`.
+- `BoardSize.isValidSquareName()`: check if a square name is valid.
 
 ### 1.2.10
 - Variants: